.claude/worktrees/serene-ishizaka/pyproject.toml
[project]
name = "kream-monitor-bot"
version = "0.1.0"
description = "크림 리셀 수익 모니터링 디스코드 봇"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.4.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "aiohttp>=3.9.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
